requirements.txt
# Phishing Detection System - Dependencies

# Core Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database & Environment
python-dotenv>=1.0.0

# Data Processing & ML
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0
xgboost>=1.7.0

# NLP & Machine Learning (Optional but recommended for full features)
# Note: torch and transformers are large; installation may take time.
torch>=1.12.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Frontend & Dashboard
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.18.0

# Security & Auth
bcrypt>=4.0.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0

# Utilities
psutil>=5.9.0
watchfiles>=0.21.0

# Cloud Database Drivers
psycopg2-binary>=2.9.9
mysql-connector-python>=8.2.0
